feat(#301) Fase 2 B1: buildBehaviorProfile + extração DRY shadowDetectors
- Extrai detectores shadow puros (analyzeShadowForTradeCF, sortChronologically,
  DEFAULT_CONFIG) de analyzeShadowBehavior.js para functions/shadow/shadowDetectors.js
  (sem firebase-functions) — desacopla p/ reuso e testabilidade. Callable mantém só
  o onCall importando do módulo puro.
- functions/behavior/buildBehaviorProfile.js: funde detectBehavior (mirror: events +
  gateInputs + scoreInputs) com shadow per-trade num behaviorProfile por trade.
  Dedupe por família reusa dedupeByFamily (DEC-074). Enriquece trades com planRoPct/
  planRrTarget (UNDERSIZED/TARGET_HIT). Fingerprint p/ idempotência (anti-loop/custo).
- 8 testes do contrato da fusão (famílias canônicas, gate flag, valência, dedupe,
  fingerprint estável). Função pura — caller decide o que gravar.
